// src/shared/types/characterDTO.ts
export interface Character {
  id: number;
  name: string;
  status: "unknown" | "Alive" | "Dead";
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}